Replace sphere character with custom GLTF model support - Added GLTFCharacter component with fallback to sphere - Integrated fish2.glb model as playable character - Added public/models directory with README instructions - Maintained all game mechanics and visual effects
